requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
black==22.1.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cycler==0.11.0
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
frozenlist==1.3.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.1.2
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.44.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.3
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.27
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
pyzmq==22.3.0
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
setuptools==60.9.3
six==1.16.0
stack-data==0.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.2
typing_extensions==4.1.1
unicodedata2==14.0.0
urllib3==1.26.8
wcwidth==0.2.5
wheel==0.37.1
xgboost==1.5.1
yarl==1.7.2
